chore: Remove old code & rename `useVinylSeekbarNext.ts` to `useVinylSeekbar.ts`
- Advantage of new implementation is that since the vinyl & regular seekbar looks at the same shared value, pausing will have both pause at the same time.
  - It also removes the janky timer logic we used for the vinyl when animating it.
