+++ mobile/src/navigation/screens/now-playing/components/Artwork.tsx
@@ -14,7 +14,7 @@ import { usePlaybackStore } from "~/stores/Playback/store";
 import { PlaybackControls } from "~/stores/Playback/actions";
 import { usePreferenceStore } from "~/stores/Preference/store";
 import { isSeekingAtom } from "../helpers/Seekbar.context";
-import { useVinylSeekbar } from "../helpers/useVinylSeekbarNext";
+import { useVinylSeekbar } from "../helpers/useVinylSeekbar";
 
 import { MediaImage } from "~/modules/media/components/MediaImage";
 import { Vinyl } from "~/modules/media/components/Vinyl";
