--- mobile/src/navigation/screens/now-playing/helpers/usePlayerProgress.ts
@@ -1,57 +1,0 @@
-import { useProgress } from "@weights-ai/react-native-track-player";
-import { atom, useAtomValue, useSetAtom } from "jotai";
-import { useMemo, useRef, useState } from "react";
-
-import { usePlaybackStore } from "~/stores/Playback/store";
-import { PlaybackControls } from "~/stores/Playback/actions";
-
-/** Dragging slider position. */
-const NextSliderPositionAtom = atom<number | null>(null);
-
-/** Safely manages the visible seekbar position to prevent "rubberbanding". */
-export function usePlayerProgress() {
-  const debounceRef = useRef<ReturnType<typeof setTimeout>>(null);
-  const [updateInterval, setUpdateInterval] = useState(200);
-  const restoredPosition = useRestorePosition();
-  const { position } = useProgress(updateInterval);
-  const nextSliderPosition = useAtomValue(NextSliderPositionAtom);
-  const setNextSliderPosition = useSetAtom(NextSliderPositionAtom);
-
-  const seekToPosition = async (progress: number) => {
-    await PlaybackControls.seekTo(progress);
-    setUpdateInterval(1);
-
-    // Helps prevents "rubberbanding".
-    if (debounceRef.current) clearTimeout(debounceRef.current);
-    debounceRef.current = setTimeout(() => {
-      setNextSliderPosition(null);
-      setUpdateInterval(200);
-    }, 200);
-  };
-
-  return {
-    position: nextSliderPosition ?? restoredPosition ?? position,
-    setPosition: setNextSliderPosition,
-    seekToPosition,
-  };
-}
-
-/** Have the seekbar visually reflect the last saved position. */
-function useRestorePosition() {
-  const _hasRestoredPosition = usePlaybackStore((s) => s._hasRestoredPosition);
-  const _restoredTrackKey = usePlaybackStore((s) => s._restoredTrackKey);
-  const activeKey = usePlaybackStore((s) => s.activeKey);
-  const lastPosition = usePlaybackStore((s) => s.lastPosition);
-
-  return useMemo(() => {
-    if (
-      _hasRestoredPosition ||
-      _restoredTrackKey === undefined ||
-      _restoredTrackKey !== activeKey
-    ) {
-      return undefined;
-    }
-    return lastPosition;
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [_hasRestoredPosition, activeKey]);
-}
